requirements.txt
Flask==2.3.2
pandas==2.0.3
scikit-learn==1.3.0
mysql-connector-python==8.0.33
joblib==1.3.2